merge_cachi2_sboms: add SBOMItem abstraction layer
An SBOM item is anything that has a name, version and purl. Such as a
CycloneDX component or SPDX package.

Will be used to implement SPDX support more seamlessly.

Also set flake8 to run on python 3.12. In Github CI, the check was
failing with a syntax error otherwise (generics syntax).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>