merge_cachi2_sboms: implement merge_spdx function
Merges the packages of both SBOMs in exactly the same way as we do for
CycloneDX.

Merges the creationInfo by simply concatenating the lists of creators of
both documents.

Merges the relationships by concatenating and processing as follows:

- if the SPDXID on either side of the relationship:
  - is the ID of the cachi2 SBOM -> replace it with the ID of the syft
    SBOM
  - is not the ID of the syft SBOM or the ID of any package in the
    merged package list -> drop the whole relationship
  - otherwise -> keep it

Note that the merging does not, in any way, mess with the "root
packages" of either document (the packages in relationship
DOCUMENT_ID DESCRIBES ROOT_ID). That means the merged document can have
two or more roots. This should be fine, especially because the
add_image_reference.py script in this repo is already able to handle
that.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>