merge_cachi2_sboms: support multi-purl packages
It's unclear at this time whether this will be needed. Syft seems to use
at most one purl for every SPDX package, but Cachi2 could conceivably
want to group multiple purls under a single package in some cases.

For example, two golang purls where the only difference is

    ?type=module vs. ?type=package

Instead of raising an error for any SPDX package with 2 or more purls,
only raise the error if the purls aren't "similar enough" (identical
except for the qualifiers).

To be extra safe, change the is_cachi2_non_registry_dependency check to
inspect all purls. That is the only piece of code that depends on the
qualifiers.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>