merge_cachi2_sboms: make test more human-friendly
The SBOM files are too large to inspect by hand. Check the property that
we care about most - the components present in the merged SBOM that were
not present in the cachi2 SBOM.

Note: this uncovers some bugs (the gomod and npm packages shouldn't be
there, they're duplicates). To be fixed later.

Also disable the flake8 line-length check; it is useless since we
already use 'black' for formatting.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>